Predicting Store Type using K Nearest Neighbors
Validating type of stores in India is always a challenging task for research and analyst company. Manually validating the store type can cost huge amount to the organizations as filed team is required to visit each and every stores of different part of the country.
Here, I have developed a simple machine learning model of K-Nearest Neighbours classifier which can predict the store type of the store. Since the KNN model is based on distance calculations between different data points, I have standardize the data to a normal scale. 
I have found that at K=3 the performance of the model is best with accuracy of 98%.
Please visit the Python file for better clarity of the model designed.

Thanks!